+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 - Bump Native SDK from v0.16.3 to v0.16.4 ([#1544](https://github.com/getsentry/sentry-unreal/pull/1544))
   - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#0164)
   - [diff](https://github.com/getsentry/sentry-native/compare/0.16.3...0.16.4)
+- Bump Java SDK from v8.53.0 to v8.54.0 ([#1551](https://github.com/getsentry/sentry-unreal/pull/1551))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#8540)
+  - [diff](https://github.com/getsentry/sentry-java/compare/8.53.0...8.54.0)
 
 ## 1.20.0